fix: address image proxy review feedback
