+++ docs/client-integration.md
@@ -46,7 +46,7 @@ curl https://your-domain.com/v1/images/generations \
   -d '{"model":"gpt-image-1","prompt":"一只戴着太空头盔的橘猫","size":"1024x1024"}'
 ```
 
-响应保持标准 `{ "created": ..., "data": [...] }` 结构，其中 `data` 项至少包含 `url` 或 `b64_json`。如果上游返回成功状态但没有图像数据，网关会将其视为异常并按渠道策略重试，避免客户端收到无法使用的空结果。
+响应保持标准 `{ "created": ..., "data": [...] }` 结构，其中 `data` 数组至少有一项包含非空的 `url` 或 `b64_json`。如果上游返回成功状态但没有图像数据，网关会将其视为异常并按渠道策略重试，避免客户端收到无法使用的空结果。
 
 ## 已验证兼容的客户端
