+++ src/server/routes/proxy/images.ts
@@ -26,6 +26,10 @@ import {
   getTesterForcedChannelId,
   selectProxyChannelForAttempt,
 } from '../../proxy-core/channelSelection.js';
+import {
+  normalizeImageGenerationRequest,
+  parseUpstreamImageResponse,
+} from '../../proxy-core/images/imageGenerationProtocol.js';
 
 export async function imagesProxyRoute(app: FastifyInstance) {
   ensureMultipartBufferParser(app);
@@ -38,7 +42,7 @@ export async function imagesProxyRoute(app: FastifyInstance) {
       });
     }
 
-    const { body, requestedModel } = input;
+    const { body, requestedModel } = input.value;
     if (!await ensureModelAllowedForDownstreamKey(request, reply, requestedModel)) return;
     const downstreamPolicy = getDownstreamRoutingPolicy(request);
     const forcedChannelId = getTesterForcedChannelId({
@@ -528,75 +532,5 @@ async function recordTokenRouterEventBestEffort(
     await operation();
   } catch (error) {
     console.warn(`[proxy/images] failed to ${label}`, error);
-  }
-}
-
-/**
- * 校验并规范化图像生成请求，避免把明显无效的请求发送到上游。
- * 图像模型的可选参数由上游决定，这里只约束 OpenAI Images API 的必填字段。
- */
-function normalizeImageGenerationRequest(body: unknown):
-  | { ok: true; body: Record<string, unknown>; requestedModel: string }
-  | { ok: false; message: string } {
-  if (!body || typeof body !== 'object' || Array.isArray(body)) {
-    return { ok: false, message: 'request body must be a JSON object' };
-  }
-
-  const normalizedBody = body as Record<string, unknown>;
-  const rawModel = normalizedBody.model;
-  if (rawModel !== undefined && typeof rawModel !== 'string') {
-    return { ok: false, message: 'model must be a string' };
-  }
-
-  const prompt = normalizedBody.prompt;
-  if (typeof prompt !== 'string' || prompt.trim().length === 0) {
-    return { ok: false, message: 'prompt is required' };
-  }
-
-  const requestedModel = (rawModel || 'gpt-image-1').trim();
-  if (!requestedModel) {
-    return { ok: false, message: 'model must not be empty' };
-  }
-
-  return {
-    ok: true,
-    body: {
-      ...normalizedBody,
-      model: requestedModel,
-      prompt: prompt.trim(),
-    },
-    requestedModel,
-  };
-}
-
-/**
- * 解析上游 Images 响应并确认至少返回一个可展示的图像结果。
- * 不符合协议的 2xx 响应会进入既有渠道重试链路，而不是静默返回空数据。
- */
-function parseUpstreamImageResponse(text: string): { ok: true; value: any } | { ok: false; message: string } {
-  try {
-    const value = JSON.parse(text) as unknown;
-    if (!value || typeof value !== 'object' || Array.isArray(value)) {
-      return { ok: false, message: 'Upstream returned an invalid image response' };
-    }
-
-    const data = (value as Record<string, unknown>).data;
-    if (!Array.isArray(data) || data.length === 0) {
-      return { ok: false, message: 'Upstream image response did not include data' };
-    }
-
-    const hasImage = data.some((item) => {
-      if (!item || typeof item !== 'object' || Array.isArray(item)) return false;
-      const record = item as Record<string, unknown>;
-      return (typeof record.url === 'string' && record.url.trim().length > 0)
-        || (typeof record.b64_json === 'string' && record.b64_json.trim().length > 0);
-    });
-    if (!hasImage) {
-      return { ok: false, message: 'Upstream image response did not include an image' };
-    }
-
-    return { ok: true, value };
-  } catch {
-    return { ok: false, message: text || 'Upstream returned malformed JSON' };
   }
 }
